feat: Add SlackChannelManagersView model
This commit adds the SlackChannelManagersView model to the beamlines.py file in the src/nsls2api/models directory. The model includes a list of slack_channel_managers and a projection setting for the corresponding database field. This model will be used to manage the Slack channel managers for each beamline.