feat: Retrieve Slack user IDs of channel managers for a beamline
This commit adds functionality to retrieve the Slack user IDs of the channel managers for a given beamline. It introduces a new function `slack_channel_managers` in the `beamline_service.py` file. The function takes the name of the beamline as input and returns a list of Slack user IDs. It also includes necessary error handling if the beamline is not found.